migrated to iced 0.8

+++ src/utility/style_constants.rs
@@ -177,10 +177,10 @@ pub const DEEP_SEA: &[u8] = include_bytes!("../../resources/palettes/DeepSea.png
 pub const MON_AMOUR: &[u8] = include_bytes!("../../resources/palettes/MonAmour.png");
 
 // font sizes
-pub const FONT_SIZE_FOOTER: u16 = 15;
-pub const FONT_SIZE_BODY: u16 = 18;
-pub const FONT_SIZE_SUBTITLE: u16 = 22;
-pub const FONT_SIZE_TITLE: u16 = 24;
+pub const FONT_SIZE_FOOTER: f32 = 15.0;
+pub const FONT_SIZE_BODY: f32 = 18.0;
+pub const FONT_SIZE_SUBTITLE: f32 = 22.0;
+pub const FONT_SIZE_TITLE: f32 = 24.0;
 
 // border styles
 pub const BORDER_WIDTH: f32 = 2.0;
